### 2019-07-09 #29 @whedon
[PRE REVIEW]: SemanticModels.jl: A Julia Package for Scientific Model Augmentation
**Submitting author:** @mehalter (<a href="http://orcid.org/0000-0002-0484-971X">Micah Halter</a>)
**Repository:** <a href="https://github.com/mehalter/juliacon2019-submission" target ="_blank">https://github.com/mehalter/juliacon2019-submission</a>

Thanks for submitting your paper to JuliaCon @mehalter. The JuliaCon editor @vchuravy, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@mehalter if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
