+++ paper/paper.yml
@@ -5,7 +5,7 @@ keywords:
   - Category Theory
 authors:
   - name: Micah Halter
-    orcid: 000-0002-0484-971X
+    orcid: 0000-0002-0484-971X
     affiliation: 1
   - name: Sreenath Raparti
     affiliation: 1
