### 2019-09-05 #29 @mehalter
Sorry for the late reply! I don't know if the reviewer has been selected yet, but maybe Chris Rackauckas. His work in partially symbolic computing with Julia fits this project very well.
