### 2019-09-23 #29 @MasonProtter
Yes, I would be interested but I don't know anything about the review process. That is, I need documentation on what is expected of me as a reviewer and what is expected of a high quality submission.
