### 2019-09-24 #29 @vchuravy
You can find information about the review process at https://proceedings.juliacon.org/guide/reviewers/ and I am also available on Slack for questions.
