### 2019-10-17 #57 @whedon
[REVIEW]: SemanticModels.jl: A Julia Package for Scientific Model Augmentation
**Submitting author:** @mehalter (<a href="http://orcid.org/0000-0002-0484-971X">Micah Halter</a>)
**Repository:** <a href="https://github.com/mehalter/juliacon2019-submission" target ="_blank">https://github.com/mehalter/juliacon2019-submission</a>

- [ ] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?

- [ ] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?
