### 2020-06-08 #57 @ChrisRackauckas
I recently took a pretty deep dive through the package and its documentation. The documentation is extremely thorough, especially things like https://aske.gtri.gatech.edu/docs/latest/examples/html/malaria.html . I have nothing to add and think this can be accepted as is. It might finally make a category theorist out of me.
