### 2020-06-08 #57 @ChrisRackauckas
The paper is mostly good. It should really be citing a lot more of the tools that it's using an some of the category theory background. Maybe a full example that's from scratch and not reading from a file? That's about it.
