### 2020-06-08 #57 @MasonProtter
I'm very sorry to say that I forgot about this, my apologies to Valentin and the submission authors.

@vchuravy It appears all the Juliacon related links are currently broken. What is the page limit for the paper? If it's more than a page, I think it'd be good if the paper was expanded a bit, but also this is fine. People can see the every extensive and well written docs / notebooks if they want to see more.

I will need a bit more time before I can say I feel that I understand the package well enough to sign off on functionality claims, but what I have seen and understand of the package so far is very impressive and interesting.

The test suite failed for me on 1.5.0-beta1, and I'm currently running it on 1.4.2 to see how that goes.

One area I'm a little unsure about, and I also don't know what the requirements / expectations are on the JuliaCon side, would be "Are there any comparable or similar packages in other programming languages?" So far as I can tell, other than saying that this package takes a different approach from Jump, Stan and ModellingToolkit it's hard for me to know how novel this work is in the wider world of scientific domain modelling.
