### 2020-07-27 #57 @vchuravy
Thank you Mason for your comments.

@mehalter thank you for your patience, can you address the reviewer comments? And is there an archival DOI for the associated code that you would like to associate with the publication?
https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance
