### 2020-07-27 #57 @mehalter
I have just tagged a new release on SemanticModels to address the notes from @MasonProtter about the test suite, this now works again in Julia v1.0-1.4.2. Julia v1.5.0-beta1 does not work, but the error comes from being unable to precompile Cassette with the new version. Given it is an unreleased version of Julia for now, we are alright with it not working quite yet until Cassette updates their package.

In regards to the paper length and adding more content, the paper was submitted as an extended abstract with a 1 page limit. As @ChrisRackauckas mentioned, with the short page limit, we were expecting to build interest in the package with the abstract and have people go check out it's documentation which we have spent a lot of time on to have full fledge examples and thorough usage notes.

Thank you to both @MasonProtter and @ChrisRackauckas for taking time to review the paper and your great feedback.

@vchuravy We do not have an archival DOI to be associated with the publication.
