### 2020-07-27 #57 @vchuravy
> @vchuravy We do not have an archival DOI to be associated with the publication.

Can you create one for SemanticsModels.jl see https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance
