### 2020-07-27 #57 @mehalter
ah sorry, also I misunderstood. We do have a DOI for SemanticModels (this is for the new v0.4.0):  10.5281/zenodo.3961696
