### 2020-07-27 #57 @whedon
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
