### 2020-07-27 #57 @mehalter
The paper PDF looks good. Thanks for your help @vchuravy
