### 2020-07-27 #57 @mehalter
@vchuravy  I just checked the page and the software repository links to the paper repository, could we get that moved to https://github.com/jpfairbanks/SemanticModels.jl
