### 2020-07-27 #57 @vchuravy
@arfon On we had the case that some folks submitted their papers in independent repositories from the actual software: As an example https://proceedings.juliacon.org/papers/10.21105/jcon.00057 points to https://github.com/mehalter/juliacon2019-submission

Is it possible to add an alias?
