### 2020-12-15 #57 @arfon
I've updated this submission to point to https://github.com/jpfairbanks/SemanticModels.jl on the JCON website.
